Add basic presence indicator
This adds the user to the `connectedUsers` list, and removes them from
the list when they leave the share or disconnect. If they are the last
to leave, the `connectedUsers` key is removed as well, making it easy
for us to see which tables have connected users.

The value of the user in the `connectedUsers` object is a timestamp, as
a future-friendly way to add a cleanup function to the database, in case
users ever fail to remove their presence.