Use permanent keys for users
This replaces the `personalDataLabel` in most applications where we are
keying off the user's name -- adding presence, adding items, and listening
to items.

Unlike the `personalDataLabel`, the `personalDataKey` is not changeable
by the user but instead is set once when the plugin first loads.

(In a future commit it will be persisted in the CODAP document.)

The only thing `personalDataLabel` is used for is for the `Name` attribute
in items.