Use allUsers list for adding listeners to data
This fixes the bug where users were not seeing data from members of a
group who had already disconnected.

Users now add themselves to both the `allUsers` and `connectedUsers`
array in Firebase.

`allUsers` takes over the roles previously used by `connectedUsers` in
the app -- it is used to list all keys where items are added, to allow
us to add listeners for each of those keys.

`connectedUsers` now only has one purpose: to alert the Firebase function
when there are no more connected users, and the table should be deleted.