Find and update user case by id on re-join
Testing: if three users join, one leaves and re-joins, their own data
should be moved to the bottom of the table, and if they specify a new
user label, their own data should be updated.

The previous search (`caseByIndex[0]`) would not reliably find the
correct case.

We need to get the id of the case first, as CODAP does not currently
support updating cases or items found by search.