Store last personal label as placeholder text
If we have this placeholder text, we don't require user to enter it
before sharing.

Once we start sharing, we have to transfer over the placeholder text to
the actual personal label, for all functions that expect to find it.

This placeholder text may be persisted to CODAP.

(Also fix clearing `joinShareId` on `leaveShare`.)