Use key+label as key for user data
This switches the key for user data in the table to be
`[permanent key]-[user label]`. This satisfies the requirement to make
it impossible for users working on other documents to be able to modify
a user's data, but also allows a user to create a new row by changing
their label, which lets users of shared documents refrain from clobbering
each other's work.

Testing: A user can re-join a shared table with the same label, and they
will be able to add to their previous work, or they can join with a new
label, and they will start a new user row.

This also cleans up the `updatePersonalDataLabelAndKey` to not return
any value, making its intent clearer.